stock_quote_alert/requirements.txt
django-gm2m==1.1.1
django-apscheduler
python-dotenv